requirements.txt
Django==4.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
gunicorn==20.1.0
django-cors-headers==3.13.0